Renamed beans name to wdk and put it in CamundaEngineConfiguration
